tienda_ropa/requirements.txt
# Django
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Base de datos
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Variables de entorno
python-dotenv==1.0.1

# Pillow para manejo de imágenes
Pillow==10.3.0

# Producción
gunicorn==22.0.0
whitenoise==6.6.0
